requirements.txt
# Core scraping library
scrapling[fetchers]>=0.2.0

# Async support
asyncio
aiohttp>=3.9.0

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0

# Optional: Redis cache
redis>=5.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Logging
loguru>=0.7.0
